Don't confuse Bundler.
When using a Bundler local-path-override for this gem, from another repo, the CWD is that of the other repo when this gemspec is evaluated.

So, we change directory to where the gemspec is to avoid confusion for the `git ls-files` files invocations.

Also:  We use null as the separator instead of newline because I'm deeply paranoid.